pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "bbearce", email = "bbearce@gmail.com"},
]
dependencies = [
    "pycrumbs @ git+https://github.com/CPBridge/pycrumbs@40390ba92e127325f7b2ebef2bc97d521334ccb3",
    "smmap==5.0.0",
    "pandas>=1.5.3",
    "python-dotenv>=1.0.0",
    "requests>=2.28.2",
    "setuptools>=67.6.0",
    "choix>=0.3.5",
    "matplotlib>=3.7.2",
    "networkx>=3.1",
    "scikit-learn>=1.3.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
